feat: write the stories in TypeScript
