+++ frontend/package.json
@@ -44,7 +44,7 @@
     "@storybook/addon-actions": "^5.3.3",
     "@storybook/addon-links": "^5.3.3",
     "@storybook/addons": "^5.3.3",
-    "@storybook/preset-create-react-app": "^1.5.1",
+    "@storybook/preset-create-react-app": "^1.5.2",
     "@storybook/react": "^5.3.3",
     "@types/react-router-dom": "^5.1.3"
   }
