+++ frontend/src/components/button/button.tsx
@@ -14,9 +14,9 @@ export enum ButtonType {
 
 export type ButtonProps = {
   type: ButtonType;
-  behavior: ButtonBehavior;
-  disabled: boolean;
-  onClick: (e: React.MouseEvent) => void;
+  behavior?: ButtonBehavior;
+  disabled?: boolean;
+  onClick?: (e: React.MouseEvent) => void;
 }
 
 export const Button: React.FC<ButtonProps> = ({type=ButtonType.Primary, behavior=ButtonBehavior.Button, disabled=false, children, onClick}) =>
