+++ frontend/src/stories/2-RangeInput.stories.tsx
@@ -8,5 +8,5 @@ export default {
 
 export const RangeInputField = () => {
   const [value, setValue] = useState(3);
-  return (<RangeInput step="1" min="0" max="5" value={value} onChange={e => setValue(e.target.value)} />);
+  return (<RangeInput step={1} min={0} max={5} value={value} onChange={e => setValue(parseInt(e.target.value, 10))} />);
 }
