+++ frontend/src/stories/3-SkillIcon.stories.tsx
