+++ frontend/src/stories/4-CloseIcon.stories.tsx
