+++ frontend/src/stories/5-LogoIcon.stories.tsx
