+++ frontend/src/stories/6-UserIcon.stories.tsx
